fixed target label updating

+++ core/src/com/beariksonstudios/existence/gameobjects/settlement/SettlementType.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.graphics.Texture;
 public interface SettlementType {
     String getName();
 
-    void update(float population);
+    void update(double population);
 
     float getGrowthRate();
